# cmd/octane/swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is the API for the Octane application.\n\nSwagger 2.0 Spec - generated by [go-swagger](https://github.com/go-swagger/go-swagger)",
    "title": "Octane",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/": {
      "get": {
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "hello"
        ],
        "summary": "Return a hello message.",
        "operationId": "HelloGET",
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          }
        }
      }
    },
    "/api/v1/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Authenticate a user.",
        "operationId": "UserLogin",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                },
                "password": {
                  "type": "string",
                  "x-go-name": "Password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    }
  },
  "responses": {
    "BadRequestResponse": {
      "description": "BadRequestResponse is a failure.",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "status_code",
          "status_message"
        ],
        "properties": {
          "message": {
            "description": "Message contains a user friendly message.",
            "type": "string",
            "x-go-name": "Message",
            "example": "The data submitted was invalid."
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 400
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "Bad Request"
          }
        }
      }
    },
    "InternalServerErrorResponse": {
      "description": "InternalServerErrorResponse is a failure.",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "status_code",
          "status_message"
        ],
        "properties": {
          "message": {
            "description": "Message contains a user friendly message.",
            "type": "string",
            "x-go-name": "Message",
            "example": "An unexpected error occurred in the application."
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 500
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "Internal Server Error"
          }
        }
      }
    },
    "LoginResponse": {
      "description": "LoginResponse returns a token.",
      "schema": {
        "type": "object",
        "required": [
          "status_code",
          "status_message"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "token"
            ],
            "properties": {
              "token": {
                "description": "Token contains the API token for authentication",
                "type": "string",
                "x-go-name": "Token",
                "example": "api-123456"
              }
            },
            "x-go-name": "Data"
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 200
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "OK"
          }
        }
      }
    },
    "NotFoundResponse": {
      "description": "NotFoundResponse is a failure.",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "status_code",
          "status_message"
        ],
        "properties": {
          "message": {
            "description": "Message contains a user friendly message.",
            "type": "string",
            "x-go-name": "Message",
            "example": "The page was not found."
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 404
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "Not Found"
          }
        }
      }
    },
    "OKResponse": {
      "description": "OKResponse is a success.",
      "schema": {
        "type": "object",
        "required": [
          "status_code",
          "status_message",
          "message"
        ],
        "properties": {
          "message": {
            "description": "Message contains a user friendly message.",
            "type": "string",
            "x-go-name": "Message",
            "example": "The operation was successful."
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 200
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "OK"
          }
        }
      }
    },
    "UnauthorizedResponse": {
      "description": "UnauthorizedResponse is a failure.",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "status_code",
          "status_message"
        ],
        "properties": {
          "message": {
            "description": "Message contains a user friendly message.",
            "type": "string",
            "x-go-name": "Message",
            "example": "You are not authorized to view this page."
          },
          "status_code": {
            "description": "Code contains the HTTP status code.",
            "type": "integer",
            "format": "int64",
            "x-go-name": "StatusCode",
            "example": 401
          },
          "status_message": {
            "description": "Status contains the string of the HTTP status.",
            "type": "string",
            "x-go-name": "StatusMessage",
            "example": "Unauthorized"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "token": {
      "description": "In the 'Value' textbox below, please enter in the word 'Bearer', a space, and then paste in your token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}